+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.4.0 on t2s2-silvio-dev
 - accept empty endDate for slots. It will internally be set to the end of the
 start day but will not be returned via JSON
+- slot json now always has creator, group and location objects instead of
+creatorId, groupId and locationId
 
 
 ## v0.3.0 on t2s2-dev